+++ CHANGELOG.md
@@ -17,5 +17,6 @@ All notable changes to this project will be documented in this file. This change
 - Add `add-annotations`, like geom_text in R.
 - Add `add-bar`, plot bar charts.
 - Add `subplot`, make subplots of existing traces.
+- Add `scatter-3d`, show 3d scatters.
 
 [Unreleased]: https://github.com/findmyway/plotly-clj